+++ source/ProcessManager.Tests/Fixtures/ProcessManagerAppFixture.cs
@@ -16,7 +16,6 @@ using System.Net.Http.Headers;
 using Azure.Core;
 using Energinet.DataHub.Core.FunctionApp.TestCommon.Azurite;
 using Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration;
-using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ResourceProvider;
 using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures;
 using Xunit.Abstractions;
 
@@ -53,11 +52,6 @@ public class ProcessManagerAppFixture : IAsyncLifetime
             appPort: 8202,
             manageDatabase: false,
             manageAzurite: false);
-
-        ServiceBusResourceProvider = new ServiceBusResourceProvider(
-            ExampleOrchestrationsAppManager.TestLogger,
-            IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace,
-            IntegrationTestConfiguration.Credential);
     }
 
     public IntegrationTestConfiguration IntegrationTestConfiguration { get; }
@@ -70,8 +64,6 @@ public class ProcessManagerAppFixture : IAsyncLifetime
 
     private AzuriteManager AzuriteManager { get; }
 
-    private ServiceBusResourceProvider ServiceBusResourceProvider { get; }
-
     public async Task InitializeAsync()
     {
         AzuriteManager.CleanupAzuriteStorage();
@@ -79,19 +71,8 @@ public class ProcessManagerAppFixture : IAsyncLifetime
 
         await DatabaseManager.CreateDatabaseAsync();
 
-        var processManagerTopicBuilder = ServiceBusResourceProvider.BuildTopic("pm-topic");
-
-        ExampleOrchestrationsAppManager.ProcessManagerTopicResources.AddSubscriptionsToTopicBuilder(processManagerTopicBuilder);
-        ProcessManagerAppManager.ProcessManagerTopicResources.AddSubscriptionsToTopicBuilder(processManagerTopicBuilder);
-
-        var processManagerTopic = await processManagerTopicBuilder.CreateAsync();
-
-        await ExampleOrchestrationsAppManager.StartAsync(
-            ExampleOrchestrationsAppManager.ProcessManagerTopicResources.CreateFromTopic(processManagerTopic),
-            ediTopicResources: null);
-
-        await ProcessManagerAppManager.StartAsync(
-            ProcessManagerAppManager.ProcessManagerTopicResources.CreateFromTopic(processManagerTopic));
+        await ExampleOrchestrationsAppManager.StartAsync(ediTopicResources: null);
+        await ProcessManagerAppManager.StartAsync();
     }
 
     public async Task DisposeAsync()
@@ -100,8 +81,6 @@ public class ProcessManagerAppFixture : IAsyncLifetime
         await ProcessManagerAppManager.DisposeAsync();
 
         await DatabaseManager.DeleteDatabaseAsync();
-
-        await ServiceBusResourceProvider.DisposeAsync();
 
         AzuriteManager.Dispose();
     }